backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai==1.0.0
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.0
python-multipart==0.0.9
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
